requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: win-64
asttokens=2.0.8=pypi_0
backcall=0.2.0=pypi_0
blas=1.0=mkl
brotli=1.0.9=h2bbff1b_7
brotli-bin=1.0.9=h2bbff1b_7
brotlipy=0.7.0=py39hb82d6ee_1004
ca-certificates=2022.07.19=haa95532_0
certifi=2022.9.24=pyhd8ed1ab_0
cffi=1.15.1=py39h0878f49_0
charset-normalizer=2.1.1=pyhd8ed1ab_0
colorama=0.4.5=pypi_0
cryptography=37.0.4=py39h7bc7c5c_0
cudatoolkit=11.6.0=hc0ea762_10
cudnn=8.4.1.50=hf5f08ae_0
cycler=0.11.0=pyhd3eb1b0_0
debugpy=1.6.3=pypi_0
decorator=5.1.1=pypi_0
docopt=0.6.2=pypi_0
entrypoints=0.4=pypi_0
executing=1.1.0=pypi_0
fonttools=4.25.0=pyhd3eb1b0_0
freetype=2.12.1=h546665d_0
icu=58.2=ha925a31_3
idna=3.4=pyhd8ed1ab_0
intel-openmp=2021.4.0=haa95532_3556
ipykernel=6.16.0=pypi_0
ipython=8.5.0=pypi_0
jedi=0.18.1=pypi_0
jpeg=9e=h8ffe710_2
jupyter-client=7.3.5=pypi_0
jupyter-core=4.11.1=pypi_0
kiwisolver=1.4.2=py39hd77b12b_0
lcms2=2.12=h2a16943_0
lerc=3.0=h0e60522_0
libbrotlicommon=1.0.9=h2bbff1b_7
libbrotlidec=1.0.9=h2bbff1b_7
libbrotlienc=1.0.9=h2bbff1b_7
libdeflate=1.12=h8ffe710_0
libpng=1.6.37=h1d00b33_4
libtiff=4.4.0=h2ed3b44_1
libuv=1.44.2=h8ffe710_0
libwebp-base=1.2.4=h8ffe710_0
libxcb=1.13=hcd874cb_1004
libzlib=1.2.12=h8ffe710_2
libzlib-wapi=1.2.12=h8ffe710_2
lz4-c=1.9.3=h8ffe710_1
m2w64-gcc-libgfortran=5.3.0=6
m2w64-gcc-libs=5.3.0=7
m2w64-gcc-libs-core=5.3.0=7
m2w64-gmp=6.1.0=2
m2w64-libwinpthread-git=5.0.0.4634.697f757=2
matplotlib=3.3.1=pypi_0
matplotlib-inline=0.1.6=pypi_0
mkl=2021.4.0=haa95532_640
mkl-include=2022.1.0=h6a75c08_874
mkl-service=2.4.0=py39h2bbff1b_0
mkl_fft=1.3.1=py39h277e83a_0
mkl_random=1.2.2=py39hf11a4ad_0
msys2-conda-epoch=20160418=1
munkres=1.1.4=py_0
nest-asyncio=1.5.5=pypi_0
numpy=1.23.3=pypi_0
openjpeg=2.5.0=hc9384bd_1
openssl=1.1.1q=h2bbff1b_0
packaging=21.3=pyhd3eb1b0_0
pandas=1.5.0=pypi_0
parso=0.8.3=pypi_0
pickleshare=0.7.5=pypi_0
pillow=9.2.0=py39hcef8f5f_2
pip=22.1.2=py39haa95532_0
pipreqs=0.4.11=pypi_0
prompt-toolkit=3.0.31=pypi_0
psutil=5.9.2=pypi_0
pthread-stubs=0.4=hcd874cb_1001
pure-eval=0.2.2=pypi_0
pycparser=2.21=pyhd8ed1ab_0
pygments=2.13.0=pypi_0
pyopenssl=22.0.0=pyhd8ed1ab_1
pyparsing=3.0.9=py39haa95532_0
pyqt=5.9.2=py39hd77b12b_6
pysocks=1.7.1=pyh0701188_6
python=3.9.13=h6244533_1
python-dateutil=2.8.2=pyhd3eb1b0_0
python_abi=3.9=2_cp39
pytorch=1.12.1=py3.9_cuda11.6_cudnn8_0
pytorch-mutex=1.0=cuda
pytz=2022.2.1=pypi_0
pywin32=304=pypi_0
pyzmq=24.0.1=pypi_0
qt=5.9.7=vc14h73c81de_0
requests=2.28.1=pyhd8ed1ab_1
setuptools=65.4.0=pypi_0
sip=4.19.13=py39hd77b12b_0
six=1.16.0=pyhd3eb1b0_1
sqlite=3.39.2=h2bbff1b_0
stack-data=0.5.1=pypi_0
tbb=2021.5.0=h2d74725_1
tk=8.6.12=h8ffe710_0
torchaudio=0.12.1=py39_cu116
torchvision=0.13.1=py39_cu116
tornado=6.2=py39h2bbff1b_0
traitlets=5.4.0=pypi_0
typing_extensions=4.3.0=pyha770c72_0
tzdata=2022c=h04d1e81_0
urllib3=1.26.11=pyhd8ed1ab_0
vc=14.2=h21ff451_1
vs2015_runtime=14.27.29016=h5e58377_2
wcwidth=0.2.5=pypi_0
wheel=0.37.1=pyhd3eb1b0_0
win_inet_pton=1.1.0=py39hcbf5309_4
wincertstore=0.2=py39haa95532_2
xorg-libxau=1.0.9=hcd874cb_0
xorg-libxdmcp=1.1.3=hcd874cb_0
xz=5.2.6=h8d14728_0
yarg=0.1.9=pypi_0
zlib=1.2.12=h8cc25b3_2
zstd=1.5.2=h6255e5f_4
